// src/renderer/types/action-types.d.ts
export const enum ActionType {
    TransitionIn = -2,
    TransitionOut = -1,
    TweenAnimation = 0,
    EffectAnimation = 1,
    ModifyVariable = 2,
    PlayAudio = 3,
    EmitEvent = 4,
    TextAnimation = 5
}

export const enum TriggerType {
    PreviousStart,
    PreviousEnd,
    Click,
    Event
}

export interface ActionBase {
    name: string;
    type: ActionType;
    trigger: TriggerType;
    event?: string;
    delay: number;
}
// ...
export interface SimpleValueTeenAnimation extends TweenAnimationBase {
    propertyName: "rotation" | "opacity";
    value: number
}

export interface ScaleTweenValueType {
    scaleX: number, scaleY: number
}

export interface ScaleTweenAnimation extends TweenAnimationBase {
    propertyName: "scale";
    value: ScaleTweenValueType;
}
// ...
export interface PositionTweenAnimation extends TweenAnimationBase {
    propertyName: "position";
    value: PositionTweenValueType;
}

export type TweenAnimation = PositionTweenAnimation | SimpleValueTeenAnimation | ScaleTweenAnimation;
// ...
export interface TextAnimation extends ActionBase {
    type: ActionType.TextAnimation;
    target: string;
    speed: number;
}

export type Action = ModifyVariableAction | TweenAnimation | PlayAudioAction | EmitEventAction | TextAnimation;

export const enum TransitionInType {
    None = 1,
    Move = 2,
    Fade = 3
}

export const enum TransitionInLayer {
    Below = 1,
    Above = 2
}

export interface TransitionInActionBase {
    type: ActionType.TransitionIn,
    transitionType: TransitionInType,
    layer: TransitionInLayer,
    duration: number;
}

export interface NoneTransitionInAction extends TransitionInActionBase {
    transitionType: TransitionInType.None;
}
// ...
export const enum TransitionOutType {
    None = 1,
    Stay = 2, // stay until transition finished
    Move = 3,
    Fade = 4
}

export type TransitionInAction = NoneTransitionInAction | FadeTransitionInAction | MoveTransitionInAction;

export interface TransitionOutActionBase {
    type: ActionType.TransitionOut;
    transitionType: TransitionOutType;
    duration: number;
}

export interface NoneTransitionOutAction extends TransitionOutActionBase {
    transitionType: TransitionOutType.None;
}

export interface DurationTransitionOutAction extends TransitionOutActionBase {
    transitionType: TransitionOutType.Stay | TransitionOutType.Fade;
}

export interface MoveTransitionOutAction extends TransitionOutActionBase {
    transitionType: TransitionOutType.Move;
    direction: TransitionMoveDirection;
}

export type TransitionOutAction = NoneTransitionOutAction | DurationTransitionOutAction | MoveTransitionOutAction;

export interface Actions {
    transitionIn: TransitionInAction;
    transitionOut: TransitionOutAction;
    actions: Action[];
}

export interface DialogueAction {
    [key: string]: Actions;
}